refactor: restructure codebase as monorepo
