+++ crates/djc-core/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "djc-core"
+version = "1.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "djc_core"
+crate-type = ["cdylib"]
+
+[dependencies]
+djc-html-transformer = { path = "../djc-html-transformer" }
+pyo3 = { workspace = true }
+quick-xml = { workspace = true }
