+++ djc_core/__init__.py
@@ -2,8 +2,8 @@
 # This file is what maturin auto-generates. But it seems maturin omits it when we have a __init__.pyi file.
 # So we have to manually include it here.
 
-from .djc_core_html_parser import *
+from .djc_core import *
 
-__doc__ = djc_core_html_parser.__doc__
-if hasattr(djc_core_html_parser, "__all__"):
-    __all__ = djc_core_html_parser.__all__
+__doc__ = djc_core.__doc__
+if hasattr(djc_core, "__all__"):
+    __all__ = djc_core.__all__
